requirements.txt
python-telegram-bot==13.15
requests~=2.28.2
loguru~=0.6.0
environs~=9.5.0